app-emulation/wine-proton: Remove dependency on GCC and MinGW when using Clang
Before this change, GCC and MinGW were dependencies even on hosts
using LLVM toolchain.

The reason behind runtime dependency on GCC was usage of libatomic. But
nowadays, compiler-rt implements atomic builtins, so the GCC libatomic
is not necessary.

GCC MinGW is not needed, because clang can serve as a drop-in
replacement. FreeBSD ports are using clang that way by default[0]. Also,
llvm-strip can be used instead of *-w64-mingw32-strip.

This change introduces the `clang` USE flag (which is enabled in LLVM
profiles) and when enabled, uses LLVM toolchain instead of GCC and MinGW.

[0] https://cgit.freebsd.org/ports/tree/emulators/wine-proton

Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>