dev-lang/rust-bin: Remove hard dependency on GCC
Achieve that by providing the `llvm-libgcc` USE flag. When enabled, it
triggers the build of llvm-libgcc[0] (libunwind with GNU symbols and a
link to compiler-rt), which is a full replacement of libgcc_s. That new
USE flag can be used on musl-llvm systems to avoid the dependency on
GCC.

This solution is temporary and we are already discussing shipping
llvm-libgcc system-wide.[1]

[0] https://github.com/llvm/llvm-project/tree/main/llvm-libgcc
[1] https://bugs.gentoo.org/946486

Bug: https://bugs.gentoo.org/944835
Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>