llvm-runtimes/libgcc: Add a drop-in replacement of libgcc_s for LLVM systems
llvm-libgcc[0] is the LLVM subproject which provides a link script,
which can be used to build a library linked to compiler-rt and
llvm-libunwind, which serves as a full replacement of libgcc_s.

The main purpose of having it in Gentoo is ability to run binaries,
linked to libgcc_s, on musl-llvm systems.

[0] https://github.com/llvm/llvm-project/tree/main/llvm-libgcc

Bug: https://bugs.gentoo.org/946486
Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>